Create logrotate
This script uses the logrotate command to rotate and compress system logs based on the configuration specified in the /etc/logrotate.conf file. The -f option is used to force the rotation of the logs even if it is not the time specified in the configuration file.

You can also create a custom logrotate configuration file for specific logs and include it in the /etc/logrotate.d/ directory, this way you can set different rotation schedule and options for different logs.

You can schedule the script to run daily, weekly, or monthly using cron, so that old logs are automatically rotated and compressed. you can use the command crontab -e to open the cronjob editor and add a schedule for the script to run.