Create Network connectivity
Yes, you can use a shell script to check the network connectivity and send an alert if the connectivity is lost. Here's an example of a simple shell script that checks the network connectivity using the ping command, and sends an email alert using the mail command:

Copy code
#!/bin/bash

# Define the IP address to ping
ip=8.8.8.8

# Send a ping request and capture the result
result=$(ping -c 1 $ip)

# Check if the ping was successful
if [[ $result == *"1 packets transmitted, 1 received"* ]]
then
    # Ping was successful, do nothing
    :
else
    # Ping was not successful, send an email alert
    echo "Network connectivity is down" | mail -s "Network Alert" your@email.com
fi
This script pings a specified IP address, in this case 8.8.8.8 (Google's DNS server) and captures the result. It then checks if the ping was successful by looking for the string "1 packets transmitted, 1 received" in the result. If the ping was not successful, the script sends an email alert using the mail command with the subject "Network Alert" to the specified email address.

You can schedule the script to run at regular intervals using cron, so that you are always aware of the network status on your system. You can use the command crontab -e to open the cronjob editor and add a schedule for the script to run.

It's worth noting that this script uses the ping command which may not be available in all system, also the command mail might not be present in all systems, in that case, you can use sendmail or mutt or other mail commands. Additionally, you might have to set up a mail server or use an external mail service to send the email.